clipboard main refactor
